refactor(admin-teardown): extract helpers to private sibling module
Move the 4 private helpers (deleteVpcConnection, waitForVpcDeletion,
cleanupIamRolePolicies, deleteIamRole) from quicksight-teardown.js into
a new private sibling quicksight-teardown-helpers.js so that
module.exports = { teardownQuickSight } is the sole public surface of
quicksight-teardown.js. Update helper tests to import from the new module.
